# docs/swagger-docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/cards": {
      "put": {
        "tags": [
          "Card-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "updateCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "card",
                  "attributes": {
                    "bankName": "ABC bank",
                    "cardNumber": "1234567890123456895",
                    "expiry": "11/31",
                    "issueDate": "11/23",
                    "cardType": "debit",
                    "cardName": "platinum"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Card-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UID for each request",
            "required": true,
            "schema": {}
          }
        ],
        "operationId": "addCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "card",
                  "attributes": {
                    "bankName": "ABC bank",
                    "cardNumber": "1234567890123456895",
                    "expiry": "11/31",
                    "issueDate": "11/23",
                    "cardType": "debit",
                    "cardName": "platinum"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Card-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {}
          }
        ],
        "operationId": "deleteCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "card",
                  "attributes": {
                    "bankName": "ABC bank",
                    "cardNumber": "1234567890123456",
                    "expiry": "11/31",
                    "issueDate": "11/23",
                    "cardType": "debit",
                    "cardName": "platinum"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cards/{cardId}": {
      "get": {
        "tags": [
          "Card-Controller"
        ],
        "operationId": "getCard",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardRequest"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "DataDTOObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "object"
          }
        }
      },
      "ErrorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "MetaDTOObject": {
        "type": "object",
        "properties": {
          "attributes": {
            "type": "object"
          }
        }
      },
      "ResponseDTOObject": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataDTOObject"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDTOObject"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
...
      "RequestId": {
        "type": "string",
        "pattern": "^[a-zA-Z0-9]",
        "description": "UUID for each request"
      },
...
      "CardRequest": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "attributes": {
                "type": "object",
                "properties": {
                  "bankName": {
                    "type": "string",
                    "pattern": "^[a-zA-Z][a-z]{1,49}$"
                  },
                  "cardNumber": {
                    "type": "string",
                    "pattern": "^[0-9]{16}$"
                  },
                  "expiry": {
                    "type": "string",
                    "format": "date"
                  },
                  "issueDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "cardType": {
                    "type": "string",
                    "pattern": "^[a-zA-Z][a-z]{1,19}$"
                  },
                  "cardName": {
                    "type": "string",
                    "pattern": "^[a-zA-Z][a-z]{1,49}$"
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}